FIX: Better logging [circle full] [skip azp] [skip actions] [skip cirrus]
