+++ mne/datasets/_fetch.py
@@ -2,7 +2,6 @@
 #
 # License: BSD Style.
 
-import logging
 import sys
 import os
 import os.path as op
@@ -18,7 +17,8 @@ from .config import (
     TESTING_VERSIONED,
     MISC_VERSIONED,
 )
-from .utils import _dataset_version, _do_path_update, _get_path
+from .utils import (_dataset_version, _do_path_update, _get_path,
+                    _log_time_size, _downloader_params)
 from ..fixes import _compare_version
 
 
@@ -222,13 +222,9 @@ def fetch_dataset(
                     "You must agree to the license to use this " "dataset"
                 )
     # downloader & processors
-    download_params = dict(progressbar=logger.level <= logging.INFO)
+    download_params = _downloader_params(auth=auth, token=token)
     if name == "fake":
         download_params["progressbar"] = False
-    if auth is not None:
-        download_params["auth"] = auth
-    if token is not None:
-        download_params["headers"] = {"Authorization": f"token {token}"}
     downloader = pooch.HTTPDownloader(**download_params)
 
     # make mappings from archive names to urls and to checksums
@@ -307,15 +303,3 @@ def fetch_dataset(
         )
     _log_time_size(t0, sz)
     return (final_path, data_version) if return_version else final_path
-
-
-def _log_time_size(t0, sz):
-    t = time.time() - t0
-    fmt = '%Ss'
-    if t > 60:
-        fmt = f'%Mm{fmt}'
-    if t > 3600:
-        fmt = f'%Hh{fmt}'
-    sz = sz / 1048576  # 1024 ** 2
-    t = time.strftime(fmt, time.gmtime(t))
-    logger.info(f'Download complete in {t} ({sz:.1f} MB)')
