+++ tools/circleci_download.sh
@@ -1,6 +1,7 @@
 #!/bin/bash -e
 
 set -o pipefail
+export MNE_TQDM=off
 
 if [ "$CIRCLE_BRANCH" == "main" ] || [[ $(cat gitlog.txt) == *"[circle full]"* ]]; then
     echo "Doing a full dev build";
